requirements.txt
django==5.1
django-cors-headers
django-allauth
django_rest_framework
python-dotenv
prettytables
djangorestframework-simplejwt
mysqlclient
